pass duplicate client test

+++ clients.py
@@ -12,13 +12,9 @@ def next_available_ip():
 def new_client(name, ip, netmask, ccd='/etc/openvpn/clients'):
     ifconfig = 'ifconfig_push {0} {1}'.format(ip, netmask)
     config = os.path.join(ccd, name)
-    print(ccd)
-    print(config)
-    print(ifconfig)
+    if os.path.isfile(config):
+        raise DuplicateClientError
     with open(config, 'w') as f:
         f.write(ifconfig)
 
     return {'name': name, 'ifconfig': ifconfig}
-
-    
-
